+++ packages/zephyr/src/Button.tsx
@@ -9,6 +9,7 @@ import classNames from 'classnames';
 import styled, { keyframes, useTheme } from 'styled-components';
 import { Box, BoxStylingProps } from './Box';
 import { ButtonVariantName } from './theme/variants/button';
+import { LeftRight, SVGComponent } from './shared';
 
 export type ButtonSize = 'extra-small' | 'small' | 'medium' | 'large';
 
@@ -22,6 +23,8 @@ export type NonSemanticButtonProps = Pick<BoxStylingProps, 'tx'> & {
    * `isLoading` can only be true for buttons with boundaries (ie. filled and outline variants). Note: `button-outline-destructive` is not currently a variant.
    */
   isLoading?: boolean;
+  adornmentLeft?: SVGComponent;
+  adornmentRight?: SVGComponent;
 };
 
 export interface ButtonProps extends PropsWithAs<'button', NonSemanticButtonProps> {}
@@ -74,9 +77,170 @@ const Loader = styled(Box)<{ shouldReduceMotion: boolean }>`
   }
 `;
 
+export type AdornmentVariation = 'leftOnly' | 'rightOnly' | 'both' | 'none';
+
+const getAdornmentVariant = (hasLeft: boolean, hasRight: boolean): AdornmentVariation => {
+  if (hasLeft && hasRight) return 'both';
+  if (hasLeft && !hasRight) return 'leftOnly';
+  if (!hasLeft && hasRight) return 'rightOnly';
+  return 'none';
+};
+
+export type ButtonCompoundVariant = `${ButtonSize}_adorned-${AdornmentVariation}`;
+
+const getCompoundVariant = (
+  size: ButtonSize | ButtonSize[],
+  adornmentVariant: AdornmentVariation,
+) =>
+  Array.isArray(size)
+    ? size.map((s) => `${s}_adorned-${adornmentVariant}` as ButtonCompoundVariant)
+    : (`${size}_adorned-${adornmentVariant}` as ButtonCompoundVariant);
+
+interface ButtonAdornmentProps extends Required<Pick<NonSemanticButtonProps, 'size'>> {
+  component: SVGComponent;
+  side: LeftRight;
+}
+
+const ButtonAdornment = ({ component, side, size }: ButtonAdornmentProps) => {
+  const theme = useTheme();
+
+  return (
+    <Box
+      as={component}
+      tx={{
+        ...styledSystemVariant({
+          prop: 'side',
+          variants: {
+            left: {
+              mr: 6,
+            },
+            right: {
+              ml: 6,
+            },
+          },
+        })({ theme, side }),
+        ...styledSystemVariant({
+          prop: 'size',
+          variants: {
+            large: {
+              width: 24,
+              height: 24,
+            },
+            medium: {
+              width: 16,
+              height: 16,
+            },
+            small: {
+              width: 16,
+              height: 16,
+            },
+            ['extra-small']: {
+              width: 16,
+              height: 16,
+            },
+          },
+        })({ theme, size }),
+      }}
+    />
+  );
+};
+
+const compoundVariantMap: { [key in ButtonCompoundVariant]: NonNullable<BoxStylingProps['tx']> } = {
+  'extra-small_adorned-none': {
+    height: 24,
+    px: 12,
+    fontSize: 12,
+  },
+  'extra-small_adorned-leftOnly': {
+    height: 24,
+    pl: 4,
+    pr: 8,
+    fontSize: 12,
+  },
+  'extra-small_adorned-rightOnly': {
+    height: 24,
+    pl: 8,
+    pr: 4,
+    fontSize: 12,
+  },
+  'extra-small_adorned-both': {
+    height: 24,
+    px: 6,
+    fontSize: 12,
+  },
+  'small_adorned-none': {
+    height: 32,
+    px: 16,
+    fontSize: 12,
+  },
+  'small_adorned-leftOnly': {
+    height: 32,
+    pl: 6,
+    pr: 10,
+    fontSize: 12,
+  },
+  'small_adorned-rightOnly': {
+    height: 32,
+    pl: 10,
+    pr: 6,
+    fontSize: 12,
+  },
+  'small_adorned-both': {
+    height: 32,
+    px: 8,
+    fontSize: 12,
+  },
+  'medium_adorned-none': {
+    height: 40,
+    px: 16,
+    fontSize: 14,
+  },
+  'medium_adorned-leftOnly': {
+    height: 40,
+    pl: 12,
+    pr: 16,
+    fontSize: 14,
+  },
+  'medium_adorned-rightOnly': {
+    height: 40,
+    pl: 16,
+    pr: 12,
+    fontSize: 14,
+  },
+  'medium_adorned-both': {
+    height: 40,
+    px: 12,
+    fontSize: 14,
+  },
+  'large_adorned-none': {
+    height: 48,
+    px: 24,
+    fontSize: 16,
+  },
+  'large_adorned-leftOnly': {
+    height: 48,
+    pl: 12,
+    pr: 20,
+    fontSize: 16,
+  },
+  'large_adorned-rightOnly': {
+    height: 48,
+    pl: 20,
+    pr: 12,
+    fontSize: 16,
+  },
+  'large_adorned-both': {
+    height: 48,
+    px: 16,
+    fontSize: 16,
+  },
+};
+
 export const Button = forwardRefWithAs<NonSemanticButtonProps, 'button'>(
   (
     {
+      adornmentLeft,
+      adornmentRight,
       as = 'button',
       disabled = false,
       size = 'medium',
@@ -104,6 +268,22 @@ export const Button = forwardRefWithAs<NonSemanticButtonProps, 'button'>(
     const theme = useTheme();
     const shouldReduceMotion = useReducedMotion();
 
+    const adornmentVariant = getAdornmentVariant(!!adornmentLeft, !!adornmentRight);
+    const compoundVariant = getCompoundVariant(size, adornmentVariant);
+
+    const content =
+      adornmentVariant === 'none' ? (
+        <>{children}</>
+      ) : (
+        <Box tx={{ display: 'inline-flex', alignItems: 'center' }}>
+          {adornmentLeft && <ButtonAdornment side="left" component={adornmentLeft} size={size} />}
+          {children}
+          {adornmentRight && (
+            <ButtonAdornment side="right" component={adornmentRight} size={size} />
+          )}
+        </Box>
+      );
+
     return (
       <Box
         as={as}
@@ -128,30 +308,9 @@ export const Button = forwardRefWithAs<NonSemanticButtonProps, 'button'>(
           lineHeight: 1,
           ...transitions(['background-color', 'color'], '0.2s ease'),
           ...styledSystemVariant({
-            prop: 'size',
-            variants: {
-              large: {
-                height: 48,
-                px: 24,
-                fontSize: 16,
-              },
-              medium: {
-                height: 40,
-                px: 16,
-                fontSize: 14,
-              },
-              small: {
-                height: 32,
-                px: 16,
-                fontSize: 12,
-              },
-              ['extra-small']: {
-                height: 24,
-                px: 12,
-                fontSize: 12,
-              },
-            },
-          })({ theme, size, ...restOfProps }),
+            prop: 'compoundVariant',
+            variants: compoundVariantMap,
+          })({ theme, compoundVariant, ...restOfProps }),
           '&:hover': {
             textDecoration: 'none',
           },
@@ -185,11 +344,11 @@ export const Button = forwardRefWithAs<NonSemanticButtonProps, 'button'>(
                 visibility: isLoading ? 'hidden' : 'visible',
               }}
             >
-              {children}
+              {content}
             </Box>
           </Box>
         ) : (
-          children
+          content
         )}
       </Box>
     );