+++ packages/zephyr/src/Menus/components/MenuDivider.tsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+import { Box, BoxStylingProps } from '../../Box';
+
+export interface MenuDividerProps extends Pick<BoxStylingProps, 'tx'> {}
+
+export const MenuDivider = ({ tx }: MenuDividerProps) => {
+  return <Box as="hr" tx={{ backgroundColor: 'pigeon100', height: 1, border: 0, my: 8, ...tx }} />;
+};