+++ packages/zephyr/src/theme/variants/text.ts
@@ -18,9 +18,13 @@ export type TextVariantName =
   | 'text-ui-20'
   | 'text-ui-18'
   | 'text-ui-16'
+  | 'text-ui-16-uppercase'
   | 'text-ui-14'
+  | 'text-ui-14-uppercase'
   | 'text-ui-12'
-  | 'text-ui-10';
+  | 'text-ui-12-uppercase'
+  | 'text-ui-10'
+  | 'text-ui-10-uppercase';
 
 const _text: { [key in TextVariantName]: TXProp } = {
   'text-heading-80': {
@@ -31,6 +35,7 @@ const _text: { [key in TextVariantName]: TXProp } = {
     fontWeight: 'medium',
     letterSpacing: '0.02em',
     lineHeight: 1.1,
+    textTransform: 'none',
   },
   'text-heading-72': {
     fontFamily: 'display',
@@ -39,6 +44,7 @@ const _text: { [key in TextVariantName]: TXProp } = {
     fontWeight: 'medium',
     letterSpacing: '0.03em',
     lineHeight: 1.1,
+    textTransform: 'none',
   },
   'text-heading-64': {
     fontFamily: 'display',
@@ -47,6 +53,7 @@ const _text: { [key in TextVariantName]: TXProp } = {
     fontWeight: 'medium',
     letterSpacing: '0.02em',
     lineHeight: 1.1,
+    textTransform: 'none',
   },
   'text-heading-56': {
     fontFamily: 'display',
@@ -55,6 +62,7 @@ const _text: { [key in TextVariantName]: TXProp } = {
     fontWeight: 'medium',
     letterSpacing: '0.03em',
     lineHeight: 1.16,
+    textTransform: 'none',
   },
   'text-heading-48': {
     fontFamily: 'display',
@@ -63,6 +71,7 @@ const _text: { [key in TextVariantName]: TXProp } = {
     fontWeight: 'medium',
     letterSpacing: '0.03em',
     lineHeight: 1.16,
+    textTransform: 'none',
   },
   'text-heading-40': {
     fontFamily: 'display',
@@ -71,6 +80,7 @@ const _text: { [key in TextVariantName]: TXProp } = {
     fontWeight: 'medium',
     letterSpacing: '0.03em',
     lineHeight: 1.15,
+    textTransform: 'none',
   },
   'text-heading-32': {
     fontFamily: 'display',
@@ -79,6 +89,7 @@ const _text: { [key in TextVariantName]: TXProp } = {
     fontWeight: 'medium',
     letterSpacing: '0.03em',
     lineHeight: 1.1,
+    textTransform: 'none',
   },
   'text-heading-24': {
     fontFamily: 'display',
@@ -87,6 +98,7 @@ const _text: { [key in TextVariantName]: TXProp } = {
     fontWeight: 'medium',
     letterSpacing: '0.03em',
     lineHeight: 1.1,
+    textTransform: 'none',
   },
   'text-heading-20': {
     fontFamily: 'display',
@@ -95,6 +107,7 @@ const _text: { [key in TextVariantName]: TXProp } = {
     fontWeight: 'medium',
     letterSpacing: '0.03em',
     lineHeight: 1.1,
+    textTransform: 'none',
   },
   'text-heading-18': {
     fontFamily: 'display',
@@ -103,6 +116,7 @@ const _text: { [key in TextVariantName]: TXProp } = {
     fontWeight: 'medium',
     letterSpacing: '0.03em',
     lineHeight: 1.1,
+    textTransform: 'none',
   },
   'text-heading-16': {
     fontFamily: 'display',
@@ -111,62 +125,111 @@ const _text: { [key in TextVariantName]: TXProp } = {
     fontWeight: 'medium',
     letterSpacing: '0.03em',
     lineHeight: 1.1,
+    textTransform: 'none',
   },
   'text-ui-36': {
     fontFamily: 'copy',
     fontFeatureSettings: `'case' on`,
     fontSize: 36,
+    letterSpacing: '-0.022em',
     lineHeight: 1.5,
+    textTransform: 'none',
   },
   'text-ui-32': {
     fontFamily: 'copy',
     fontFeatureSettings: `'case' on`,
     fontSize: 32,
     lineHeight: 1.5,
-    letterSpacing: '0.001em',
+    letterSpacing: '-0.021em',
+    textTransform: 'none',
   },
   'text-ui-24': {
     fontFamily: 'copy',
     fontFeatureSettings: `'case' on`,
     fontSize: 24,
+    letterSpacing: '-0.0019em',
     lineHeight: 1.2,
+    textTransform: 'none',
   },
   'text-ui-20': {
     fontFamily: 'copy',
     fontFeatureSettings: `'case' on`,
     fontSize: 20,
+    letterSpacing: '-0.017em',
     lineHeight: 1.5,
-    letterSpacing: '-0.03em',
+    textTransform: 'none',
   },
   'text-ui-18': {
     fontFamily: 'copy',
     fontFeatureSettings: `'case' on`,
     fontSize: 18,
+    letterSpacing: '-0.014em',
     lineHeight: 1.5,
-    letterSpacing: '-0.024em',
+    textTransform: 'none',
   },
   'text-ui-16': {
     fontFamily: 'copy',
     fontFeatureSettings: `'case' on`,
     fontSize: 16,
+    letterSpacing: '-0.011em',
     lineHeight: 1.5,
-    letterSpacing: '-0.02em',
+    textTransform: 'none',
+  },
+  'text-ui-16-uppercase': {
+    fontFamily: 'copy',
+    fontFeatureSettings: `'pnum' on, 'lnum' on, 'ss01' on`,
+    fontSize: 16,
+    letterSpacing: '0.1em',
+    lineHeight: 1.5,
+    textTransform: 'uppercase',
   },
   'text-ui-14': {
     fontFamily: 'copy',
+    fontFeatureSettings: 'normal',
     fontSize: 14,
-    letterSpacing: '-0.015em',
+    letterSpacing: '-0.006em',
     lineHeight: 1.5,
+    textTransform: 'none',
+  },
+  'text-ui-14-uppercase': {
+    fontFamily: 'copy',
+    fontFeatureSettings: `'cpsp' on, 'case' on`,
+    fontSize: 14,
+    letterSpacing: '0.04em',
+    lineHeight: 1.5,
+    textTransform: 'uppercase',
   },
   'text-ui-12': {
     fontFamily: 'copy',
+    fontFeatureSettings: 'normal',
     fontSize: 12,
+    letterSpacing: 'normal',
     lineHeight: 1.5,
+    textTransform: 'none',
+  },
+  'text-ui-12-uppercase': {
+    fontFamily: 'copy',
+    fontFeatureSettings: `'pnum' on, 'lnum' on, 'ss01' on`,
+    fontSize: 12,
+    letterSpacing: '0.03em',
+    lineHeight: 1.5,
+    textTransform: 'uppercase',
   },
   'text-ui-10': {
     fontFamily: 'copy',
+    fontFeatureSettings: 'normal',
     fontSize: 10,
+    letterSpacing: '0.01em',
     lineHeight: '12px',
+    textTransform: 'none',
+  },
+  'text-ui-10-uppercase': {
+    fontFamily: 'copy',
+    fontFeatureSettings: `'pnum' on, 'lnum' on, 'ss01' on`,
+    fontSize: 10,
+    letterSpacing: '0.04em',
+    lineHeight: '12px',
+    textTransform: 'uppercase',
   },
 };